cuando ingreso bye ya termino a los dos hijos me falta hacer que termine el padre

+++ tp6/mmap_sig.py
@@ -1,4 +1,4 @@
-import getopt, sys, os, mmap, signal, time
+import getopt, sys, os, mmap, signal
 # try:
 #     opt,arg = getopt.getopt(sys.argv[1:], 'f:')
 #     if len(opt) != 1:
@@ -12,22 +12,35 @@ import getopt, sys, os, mmap, signal, time
 #     if op == '-f':
     #    path = str(ar)
 
-# fd_bajo_nievl = os.open(path, os.O_RDWR) # lo abro en bajo nivel y en lectura y escritura
+# path = '/tmp/archivo.txt'
+path = 'file.log'
 
 memoria = mmap.mmap(-1, 100)
 
 def handler_padre(s, f):
-    linea = memoria.readline()
-    print(f'El padre acaba de recibir desde H1: {linea.decode()}')
-    
-    os.kill(pidh2, signal.SIGUSR1) # enviar
+    # print(s == signal.SIGUSR1, '¿es la señal SIGUSR1?') # con esto??
+    if s == signal.SIGUSR1:
+        linea = memoria.readline()
+        print(f'El padre acaba de recibir desde H1: {linea.decode()}')
+        os.kill(pidh2, signal.SIGUSR1) # enviar
+    if s == signal.SIGUSR2:
+        print('El padre le avisar a h2 que tiene que terminar')
+        os.kill(pidh2, signal.SIGUSR2)
+        # os._exit()
+
+
 
 def handler_h2(s, f):
-    linea = memoria.readline()
-    print(f'El H2 acaba de recibir la señal del padre y lee la linea: {linea.decode()}')
+    if s == signal.SIGUSR1:
+        linea = memoria.readline()
+        print(f'El H2 acaba de recibir la señal del padre y lee la linea: {linea.decode()}')
 
-    # with open(,¡....º) as archivo:
-    #     archivo.write(....)
+        with open(path, 'a') as archivo:
+            archivo.write(linea.decode())
+            archivo.flush()
+    if s == signal.SIGUSR2:
+        print('H2 muriendo y avisando al padre')
+        os._exit(0)
 
 pidh1 = os.fork()
 if pidh1 == 0:
@@ -35,18 +48,23 @@ if pidh1 == 0:
     print(f'Soy H1 mi pid es: {pidh1} y mi ppid es: {os.getppid()}\n')
     
     for linea in sys.stdin:
-        memoria.write(linea.encode('ascii'))
-        os.kill(os.getppid(), signal.SIGUSR1)
+        # print(linea == 'bye\n', 'Es igual a bye??')
+        if linea == 'bye\n':
+            print('H1 muriendo y avisando al padre!')
+            os.kill(os.getppid(), signal.SIGUSR2)
+            # os._exit(0)
+        else:
+            print(f'H1 acaba de recibir la linea: {linea}')
+            memoria.write(linea.encode('ascii'))
+            os.kill(os.getppid(), signal.SIGUSR1)
 
-    os._exit(0)
+    # os._exit(0)
 
-# pid_h2 = 0
-# esto es el padre
 pidh2 = os.fork()
 if pidh2 == 0:
     print(f'Soy H2 y mi pid es: {pidh2} y mi ppdid es: {os.getppid()}\n')
-    # pid_h2  = os.getpid()
     signal.signal(signal.SIGUSR1, handler_h2)
+    signal.signal(signal.SIGUSR2, handler_h2)
     while True:
         signal.pause()
 
@@ -54,9 +72,16 @@ if pidh2 == 0:
 
 # esto es el padre
 print(f'Soy el padre y mi pid es: {os.getpid()}\n')
+# signal.signal(signal.SIGUSR2, handler_padre)
 signal.signal(signal.SIGUSR1, handler_padre)
+signal.signal(signal.SIGUSR2, handler_padre)
 while True:
     signal.pause()
+
+
+# else:
+#     for i in range(2):
+#         os.wait()
 # os.wait()
